+++ src/remote-server/startRemoteServer.ts
@@ -6,8 +6,7 @@ import { spaceTrim } from 'spacetrim';
 import swaggerJsdoc from 'swagger-jsdoc';
 import swaggerUi from 'swagger-ui-express';
 import { forTime } from 'waitasecond';
-import { CLAIM } from '../config';
-import { DEFAULT_IS_VERBOSE } from '../config';
+import { CLAIM, DEFAULT_IS_VERBOSE } from '../config';
 import { assertsError } from '../errors/assertsError';
 import { AuthenticationError } from '../errors/AuthenticationError';
 import { PipelineExecutionError } from '../errors/PipelineExecutionError';
@@ -24,14 +23,12 @@ import { $provideFilesystemForNode } from '../scrapers/_common/register/$provide
 import { $provideScrapersForNode } from '../scrapers/_common/register/$provideScrapersForNode';
 import { $provideScriptingForNode } from '../scrapers/_common/register/$provideScriptingForNode';
 import { promptbookFetch } from '../scrapers/_common/utils/promptbookFetch';
-import type { InputParameters } from '../types/typeAliases';
-import type { string_pipeline_url } from '../types/typeAliases';
+import type { InputParameters, string_pipeline_url } from '../types/typeAliases';
 import { keepTypeImported } from '../utils/organization/keepTypeImported';
 import type { really_any } from '../utils/organization/really_any';
 import type { TODO_any } from '../utils/organization/TODO_any';
 import type { TODO_narrow } from '../utils/organization/TODO_narrow';
-import { BOOK_LANGUAGE_VERSION } from '../version';
-import { PROMPTBOOK_ENGINE_VERSION } from '../version';
+import { BOOK_LANGUAGE_VERSION, PROMPTBOOK_ENGINE_VERSION } from '../version';
 import type { RemoteServer } from './RemoteServer';
 import type { PromptbookServer_Error } from './socket-types/_common/PromptbookServer_Error';
 import type { Identification } from './socket-types/_subtypes/Identification';
@@ -41,8 +38,7 @@ import type { PromptbookServer_PreparePipeline_Request } from './socket-types/pr
 import type { PromptbookServer_PreparePipeline_Response } from './socket-types/prepare/PromptbookServer_PreparePipeline_Response';
 import type { PromptbookServer_Prompt_Request } from './socket-types/prompt/PromptbookServer_Prompt_Request';
 import type { PromptbookServer_Prompt_Response } from './socket-types/prompt/PromptbookServer_Prompt_Response';
-import type { LoginResponse } from './types/RemoteServerOptions';
-import type { RemoteServerOptions } from './types/RemoteServerOptions';
+import type { LoginResponse, RemoteServerOptions } from './types/RemoteServerOptions';
 
 keepTypeImported<PromptbookServer_Prompt_Response>(); // <- Note: [🤛]
 keepTypeImported<PromptbookServer_Error>(); // <- Note: [🤛]
@@ -77,25 +73,6 @@ export function startRemoteServer<TCustomOptions = undefined>(
         login: null,
         ...options,
     };
-    // <- TODO: [🦪] Some helper type to be able to use discriminant union types with destructuring
-    let { rootPath = '/' } = options;
-
-    if (!rootPath.startsWith('/')) {
-        rootPath = `/${rootPath}`;
-    } /* not else */
-    if (rootPath.endsWith('/')) {
-        rootPath = rootPath.slice(0, -1);
-    } /* not else */
-    if (rootPath === '/') {
-        rootPath = '';
-    }
-
-    const socketioPath =
-        '/' +
-        `${rootPath}/socket.io`
-            .split('/')
-            .filter((part) => part !== '')
-            .join('/');
 
     const startupDate = new Date();
 
@@ -177,17 +154,22 @@ export function startRemoteServer<TCustomOptions = undefined>(
                 version: '1.0.0',
                 description: 'API documentation for the Promptbook Remote Server',
             },
+            /*
+            TODO:
             servers: [
                 {
                     url: `http://localhost:${port}${rootPath}`,
-                    // <- TODO: !!!!! Probbably: Pass `remoteServerUrl` instead of `port` and `rootPath`
+                    // <- TODO: Pass some public URLs here
                 },
             ],
+            */
         },
         apis: ['./src/remote-server/**/*.ts'], // Adjust path as needed
     };
 
     const swaggerSpec = swaggerJsdoc(swaggerOptions);
+
+    const rootPath = ''; // <- TODO: !!!! Remove after merging into feature/elysia+openai+swagger-2
 
     app.use([`/api-docs`, `${rootPath}/api-docs`], swaggerUi.serve, swaggerUi.setup(swaggerSpec));
 
@@ -620,11 +602,12 @@ export function startRemoteServer<TCustomOptions = undefined>(
     const httpServer = http.createServer(app);
 
     const server: Server = new Server(httpServer, {
-        path: socketioPath,
+        path: '/socket.io',
         transports: [/*'websocket', <- TODO: [🌬] Make websocket transport work */ 'polling'],
         cors: {
             origin: '*',
             methods: ['GET', 'POST'],
+            // <- TODO: [🌡] Allow to pass
         },
     });
 
@@ -822,7 +805,7 @@ export function startRemoteServer<TCustomOptions = undefined>(
 }
 
 /**
- * TODO: !! Add CORS and security - probbably via `helmet`
+ * TODO: [🌡] Add CORS and security - probbably via `helmet`
  * TODO: Split this file into multiple functions - handler for each request
  * TODO: Maybe use `$exportJson`
  * TODO: [🧠][🛍] Maybe not `isAnonymous: boolean` BUT `mode: 'ANONYMOUS'|'COLLECTION'`